moved essential.dart file to new asset directory
