+++ lib/src/screens/card_screen.dart
@@ -4,7 +4,7 @@
 // This is a simple framework card screen.
 //
 
-import '../widgets/essential.dart';
+import '../assets/essential.dart';
 import '../widgets/card.dart';
 import 'dart:math';
